move translatables to separate file
